refactor(app): extract useDirtyState hook (zif)
Moves dirtyBlockIds, dirtyEditors, their ref mirrors, hasUnsavedSettings,
and saveStatus out of App.tsx into src/hooks/useDirtyState.ts. The stable
callback refs for handleSaveAll/handleSaveProjectSettings remain in App.tsx
due to a circular dependency with useProjectIO.
